pyproject.toml
[project]
name = "text2sql-hiring-demo"
version = "0.1.0"
description = "问数机器人-测试（马上赢）"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pyyaml>=6.0.3",
]


[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.scripts]
start = "text2sql_hiring_demo.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v -s"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
